Added qtutils; a bunch of utilities for facilitating multithreaded programming in qt. They are: qtlock: a context manager which ensures the mainloop is blocked whilst it is held. inmain(fn,*args,**kwargs): runs a function in the mainloop, waiting on it and returning its result. inmain_later(fn,*args,**kwargs): ask for a function to be run in the mainloop, not waiting on it. inmain_decorator(wait=True): a decorator which causes decorated functions to be executed in the mainloop (and either waited on on to return or not).
Qt and Python are very hackable, and thanks to this the above utilities have functionality built in to protect from their misuse. Any deadlocks arising from their misuse ought to be raised as an exception, and any Qt method calls that aren't done in a threadsafe way ought to raise an exception too. Further, Qt normally spouts (contextless, impossible to see the origin of) warnings when threads are used incorrectly, these have been promoted to exceptions with full tracebacks.