Made a toolpalettegroup class. This now puts hide/show buttons at the top of each ToolPalette You can also link the widths of items across multiple toolpalettes so that they look nice! Fixed a bug with the 'number of widgets in a row' calculation. This still needs a bit of work as there is a fudge factor in there still.
Still need to implement an icon in the hide/show button to indicate the hide/show state (like an arrow)
Need to implement some reorder methods for the toolpalettegroup.
Need to implement remove methods for the toolpalettegroup and the linked width group
Possibly should replace my complex dictionary/tuple structure with classes
Need to work out why the 'tab key' ordering is strange

Would be cool if I could implement drag/drop for the toolpalletes with the toolpalette group!