Added some code (that I got from the internet, it is licensed under the MIT license) that modifies the behaviour of the QUiLoader
This code allows you to create the "QMainWindow" in Qt Designer, but override the QMainWindow class with your own when the UI is instantiated. This means you can subclass QMainWindow and override functions like closeEvent, while still designing the window in Qt Designer. See the BLACSWindow class in pythonlib/BLACS/main.pyw for an example implementation