Rewrote the UiLoader class to be far better than the code I found on the internet
You can now promote widgets at runtime!
It is now much more general, and can apply to widgets beyond just the toplevel widget

As such there is now no trace of the code from the internet, so I'm relicensing this under the two-clause BSD as the rest of our work is.