Fixed as far as I know all corner cases for the outputbox.
It is up to callers to provide a newline character or not - outputbox will not
add one in. It should faithfully reproduce what it is sent, with unicode, colour
for stderr, and removing lines after it has more than 1000 (settable at
instantiation). The remaining scrolling issues seem fixed once and for all,
and there should be an appropriate font on every platform.