Added auto_scroll_to_end module, with its method set_auto_scroll_to_end
This is a function that takes a scrollbar as an argument, and configures it
to be kept scrolled to the end whenever the range of the scrollbar increases,
but only if it was already scrolled to the end before the range was increased.

This is useful for text output and adding items to a list.

Modified OutputBox to use this instead of duplicating code.

Bumped patch version number.