PySide2 support
PySide2 looks to be the future of Qt in Python so I figured we should add support,
it was not too hard.

The version of PySide2 in conda-forge at the moment is an alpha and has some
missing widgets and raises NotImplementedErrors on a few things, so labscript programs
are not working perfectly. But these are issues with PySide2, not qtutils.

The only issue I've seen that might be the a problem with qtutils so far is that the outputbox
text isn't white. Not sure why.

There are other rough edges - some PySide2 methods appear to not accept say, integers,
requiring the right Qt datatype such as a Qt.Checkstate. So where integers have
been saved to config files, when restored they need to be wrapped in Qt.CheckState(int) before being
used in Qt methods. And they will need to converted to integers before being saved to the config file
so that they are compatible betwen Qt wrappers. This means labscript programs are all erroring on
loading their config files.

Either PySide2 will sort issues like this out (it seems to be in very active development at the moment)
or fixes can be applied to the applications. There is plenty of time, but basic
qtutils support seems to be step 1 so here it is.

For now PySide2 is listed as the lowest priority Qt lib in qtutils.qt, so it
is unlikely anyone will have issues for the moment, but I expect it will be more
widely used in the future and probably become the default Qt in
Anaconda.