Make outputbox more responsive for large numbers of lines.
The time to print lines to the GUI scales with the number of lines, and for only
modest numbers of lines, the processing time in the GUI thread exceeds the overhead
of posting the function call to the main thread. Therefore one might as well
make smaller batches, and update the GUI more frequently to not hang the GUI in
one large call and so that output is continuously produced rather than all at
once after a hang. A batch size as small as ten seems to have good performance.

using inmain_decorator(False) also seems to make output faster.

This change actually seems to make output faster all up, i.e. more throughput,
not just better responsiveness. I'm not quite sure why, but it is a nice improvement.