Restructure into our new standard package layout
Remove compiled icons files and incorporate building them as a build
step in setup.py.

Use setuptools_scm for versioning during packaging/development and
importlib_metadata at runtime.

Remove building of PyQt4 and PySide icons

This commit does not actually move the icon files and so is
non-functional by itself - this is done to keep the diff for this branch
readable. This branch is functional if merged with the 'move-icons'
branch.