Do not include run_exports from pyqt as a run-time dependency
the pyqt conda packages lists itself under its `run_exports`, which
means that it is automatically considered a run-time dependency of any
package of which it is a build dependency. Moreover, the run-time
dependency is pinned to the specific version of PyQt used during the
build, which would be excessive even if it were a run-time requirement.

setuptools-conda 0.6.7 adds an argument for passing a list of packages
to the ignore_run_exports section of the recipe, so we use that to
ensure pyqt is not a run-time dependency.

Also, due to some strangeness on conda-forge (required for PySide2
during the build), the python_abi of the specific Python package used
during the build is being included as a run-time dependency. Ignore it
as well. It sounds like this can be removed once a newer version of
conda-build is in the conda repositories that resolves the issue.